bugfix: phantom button_up messages
I stupidly assumed the wacom driver did not send MK_LBUTTON messages and
did my own hack to detect when the pen lost contact with the tablet.
This worked on most devices as long as they always reported pressure data
while the pen was being dragged. This was not the case on at least one
device (Intuos 5 Small).