[bugfix] dynamic array reserve() bug
Was setting the new array capacity before the if block, in which it
checks if it should grow the array. The result was that arrays that were
reserved for a newer, bigger size were not being allocated. It was
relatively rare since this is only used for strokes and history, but it
did happen when opening files after having loaded a smaller one.