Refined the linux build process a bit.
Introduced a build-lin.sh bash script that does the following:
1. Builds cmake and installs its static libraries and linux-specific headers
    into the third_party/SDL2-2.0.3/build/linux64/{lib,include} directories.
2. Build Milton, linking with the SDL libs that were created in step 1.

Prior to this commit, cmake would try to use the system-installed SDL2, which
has a couple of annoyances:
1. It's not uncommon for a system not to have SDL2, so the user has to install it.
   It might be a different version than the one that Milton was written to work with,
   potentially causing difficult-to-track-down bugs.
2. There was not a FindSDL2.cmake module packaged along with Milton, which could mean
   an irritating process of trying to get cmake to find where the SDL2 libraries and
   headers live on a system. This is actually what bit me when I tried to build the
   linux version.

Finaly, Milton already includes a copy of the SDL2 library in its repo which
it uses for windows builds, so we should just use the included copy on linux too!
:)

There are also some other smaller improvements such as:
* Using a variable for the third_party/sdl2 directory so it isn't repeated a lot in CMakeLists.txt
* Ensuring CMakeLists.txt require libpthread and libdl on linux linker errors would happen otherwise.

For the future:
The cmake command used to build SDL2 inside of the build-lin.sh script should probably be made into
a 'ExternalProject_Add' statement inside the CMakelists.txt file, and used across all platforms this way
to build sdl2, and overall the amount of `if(linux) x else y` situations in the cmake script should be
reduced greatly.
This ExternalProject_Add would also replace the current add_subdirectory() statement in CMakeLists.txt.
The reason I think this is necessary is that on linux, the install process of the SDL2 project actually
needs to be executed in order to produce header files that work on linux (contain the necessary X11 content);
It's apparently not sufficient to just use third_party/SDL2-2.0.3/include directory like you can on windows.