Fix #140 add Cancel option to new canvas dialog
When creating a new canvas the old progress might be lost. The user is
queried with a dialog to save their progress. A new option "cancel" is
now available which lets you continue working on the current canvas
instead of creating a new one.

Also, cancelling the save dialog in case you answer "yes" to saving
your progress will also get you back to the current canvas instead of
creating a new one.