Add command line build-and-run script
My workflow is to edit code with a text editor and then compile and run
Milton from the command line. These are the tools I use for it.

The link (.lnk) to Visual Studio 2019's default location of the x64 dev
cmd should work on other PCs as well, as long as VS19 is installed in
its default directory.

The run.bat script will build and run Milton, stopping on compile
errors and printing Milton's stdout output after running it. This makes
it easy to use printf-debugging.